WIP: Notifications, storage, fixes

+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/service/devices/vivomovehr/notifications/NotificationStorage.java
@@ -1,31 +1,40 @@
 package nodomain.freeyourgadget.gadgetbridge.service.devices.vivomovehr.notifications;
 
 import android.util.SparseArray;
+import android.util.SparseIntArray;
 import android.util.SparseLongArray;
-import nodomain.freeyourgadget.gadgetbridge.model.NotificationSpec;
 import nodomain.freeyourgadget.gadgetbridge.service.devices.vivomovehr.ancs.AncsCategory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.HashSet;
 import java.util.LinkedHashMap;
+import java.util.Map;
 import java.util.Set;
 
 public class NotificationStorage {
+    private static final Logger LOG = LoggerFactory.getLogger(NotificationStorage.class);
+
     private static final long EXPIRATION = 30 * 1000;
 
     private final Object lock = new Object();
-    private final SparseArray<NotificationSpec> storage = new SparseArray<>();
+    private final SparseArray<NotificationData> storage = new SparseArray<>();
     private final LinkedHashMap<Long, Set<Integer>> expirationQueue = new LinkedHashMap<>();
     private final SparseLongArray expirationForNotification = new SparseLongArray();
 
-    public void registerNewNotification(NotificationSpec notificationSpec, AncsCategory category) {
+    private final SparseIntArray categoryCounts = new SparseIntArray(AncsCategory.values().length);
+
+    public void registerNewNotification(NotificationData notificationData) {
         final long now = System.currentTimeMillis();
         final long expiration = now + EXPIRATION;
 
-        final int notificationId = notificationSpec.getId();
+        final int notificationId = notificationData.spec.getId();
 
         synchronized (lock) {
             cleanup();
-            storage.put(notificationId, notificationSpec);
+            storage.put(notificationId, notificationData);
+            final int category = notificationData.category.ordinal();
+            categoryCounts.put(category, categoryCounts.get(category) + 1);
             expirationForNotification.put(notificationId, expiration);
 
             Set<Integer> expirationSet = expirationQueue.get(expiration);
@@ -39,7 +48,12 @@ public class NotificationStorage {
 
     public void deleteNotification(int id) {
         synchronized (lock) {
-            storage.delete(id);
+            final NotificationData notificationData = storage.get(id);
+            if (notificationData != null) {
+                storage.delete(id);
+                final int categoryOrdinal = notificationData.category.ordinal();
+                categoryCounts.put(categoryOrdinal, categoryCounts.get(categoryOrdinal) - 1);
+            }
             final long expiration = expirationForNotification.get(id);
             final Set<Integer> expirationSet = expirationQueue.get(expiration);
             if (expirationSet != null) {
@@ -49,7 +63,7 @@ public class NotificationStorage {
         }
     }
 
-    public NotificationSpec retrieveNotification(int id) {
+    public NotificationData retrieveNotification(int id) {
         synchronized (lock) {
             cleanup();
             return storage.get(id);
@@ -57,10 +71,31 @@ public class NotificationStorage {
     }
 
     public int getCountInCategory(AncsCategory category) {
-        return 1;
+        synchronized (lock) {
+            return categoryCounts.get(category.ordinal());
+        }
     }
 
     private void cleanup() {
+        final long now = System.currentTimeMillis();
+        Set<Integer> expiredNotifications = null;
+        for (final Map.Entry<Long, Set<Integer>> entry : expirationQueue.entrySet()) {
+            final long expiration = entry.getKey();
+            if (expiration > now) break;
 
+            final Set<Integer> setToExpire = entry.getValue();
+            if (expiredNotifications == null) expiredNotifications = new HashSet<>(setToExpire.size());
+
+            expiredNotifications.addAll(setToExpire);
+        }
+
+        if (LOG.isDebugEnabled()) {
+            LOG.debug("{}/{}/{} notification(s) in storage, removing {}", storage.size(), expirationQueue.size(), expirationForNotification.size(), expiredNotifications == null ? 0 : expiredNotifications.size());
+        }
+        if (expiredNotifications == null) return;
+
+        for (final Integer toExpire : expiredNotifications) {
+            deleteNotification(toExpire);
+        }
     }
 }
